3D viewer: use buffer geometry for label spheres
By default the 3D viewer fetches labels for all visible skeletons and
displays them as small spheres. In the past, for every label one
THREE.Mesh instance was created. The same is true for connector handles.
With many labels/connectors or many skeletons this becomes a problem
quickly. This is noticeable especially when removing objects from a
scene.

Using one buffer geometry for all labels in a skeleton fixes this at
least for labels. It reduces the number of mesh instances considerably
(of course only if labels are present). The next step will be be to have
a similar implementation for connector spheres.

See catmaid/CATMAID#1210